trying to do the thing bel told me IDK
