+++ src/app/app.component.ts
@@ -1,10 +1,13 @@
 import { Component } from '@angular/core';
-import {MatFormFieldModule} from '@angular/material/form-field';
 import { RouterOutlet } from '@angular/router';
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule} from '@angular/material/icon';
+import {MatToolbarModule} from '@angular/material/toolbar';
+import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'app-root',
-  imports: [MatFormFieldModule, RouterOutlet],
+  imports: [RouterOutlet, MatButtonModule, MatIconModule, MatToolbarModule, RouterLink],
   standalone: true,
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
