+++ src/app/app.routes.ts
@@ -2,8 +2,10 @@ import { RecipesComponent } from './recipes/recipes.component';
 import { RecipeViewComponent } from './recipes/recipe-view/recipe-view.component';
 import { Routes } from '@angular/router'
 import { RecipeEditComponent } from './recipes/recipe-edit/recipe-edit.component';
+import { CreateRecipeComponent } from './recipes/create-recipe/create-recipe.component';
 
-export const routes: Routes = [
+
+  export const routes: Routes = [
   {
     path: '',
     component: RecipesComponent
@@ -15,6 +17,10 @@ export const routes: Routes = [
   {
     path: 'recipe-edit-component/:id',
     component: RecipeEditComponent
+  },
+  {
+    path: 'recipe-create',
+    component: CreateRecipeComponent
   }
 
 ];
