+++ src/app/recipes/create-recipe/create-recipe.component.css
@@ -1,0 +1,13 @@
+.float-parent-element {
+  width: 100%;
+  display: flex;
+  gap: 16px;
+}
+.float-child-element {
+  float: left;
+  width: 50%;
+}
+
+ul {
+  list-style-type: none;
+}
