+++ src/app/recipes/create-recipe/create-recipe.component.html
@@ -1,0 +1,2 @@
+<app-new-recipe (addNewRecipe)="onAddRecipeTest()"class="float-child-element" />
+<app-ingredients [ingredientsList]="ingredients" class="float-child-element" />
