+++ src/app/recipes/create-recipe/ingredients/ingredients.component.css
@@ -4,16 +4,6 @@
   align-items: flex-start;
 }
 
-
-form {
-  margin: 2rem auto;
-  max-width: 30rem;
-  padding: 2rem;
-  border-radius: 4px;
-  background-color: #44433f;
-  box-shadow: 0 0 8px slategrey;
-}
-
 input,
 textarea,
 mat-form-field {
