+++ src/app/recipes/create-recipe/ingredients/ingredients.model.ts
@@ -1,4 +1,5 @@
 export interface Ingredients {
+  id: string;
   name: string;
   quanity: string;
   unit: 'tablespoon' | 'teaspoon' | 'cup' | 'kilogram' | 'gram';
