+++ src/app/recipes/create-recipe/new-recipe/new-recipe.component.css
@@ -4,14 +4,14 @@
   align-items: flex-start;
 }
 
-
-div {
-  margin: 2rem auto;
-  max-width: 30rem;
-  padding: 2rem;
-  border-radius: 4px;
-  background-color: #44433f;
-  box-shadow: 0 0 8px slategrey;
+input,
+textarea,
+mat-form-field {
+  font: inherit;
+  width: 100%;
+  padding: 0.5rem;
+  margin-bottom: 0.25rem;
+  border: none;
 }
 
 #actions {
@@ -31,8 +31,4 @@ button {
 button:hover,
 button:active {
   background-color: #f3d861;
-}
-
-ul {
-  list-style-type: none;
 }
