+++ src/app/recipes/create-recipe/new-recipe/new-recipe.component.ts
@@ -1,7 +1,7 @@
 import { type Ingredients } from './../ingredients/ingredients.model';
 import { Component, inject, output } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { RecipesService } from '../recipes.service';
+import { RecipesService } from '../../recipes.service';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule} from '@angular/material/input'
 import { MatSelectModule} from '@angular/material/select';
@@ -17,13 +17,11 @@ export class NewRecipeComponent {
   enteredTitle = '';
   enteredIngredients = '';
   enteredProcess = '';
-  enteredUnit: 'tablespoon' | 'teaspoon' | 'cup' | 'kilogram' | 'gram'= 'tablespoon';
+  enteredUnit: 'tablespoon' | 'teaspoon' | 'cup' | 'kilogram' | 'gram' = 'tablespoon';
   enteredQuanity = '';
   addNewRecipe = output<void>();
-  ingredient = output<Ingredients>();
 
   private recipesService = inject(RecipesService)
-
 
   onSubmit(){
     this.recipesService.addRecipe(this.enteredTitle, this.enteredProcess);
@@ -33,16 +31,4 @@ export class NewRecipeComponent {
     this.addNewRecipe.emit();
   }
 
-  onAddIngredient(){
-    this.ingredient.emit({
-      name: this.enteredIngredients,
-      quanity: this.enteredQuanity,
-      unit: this.enteredUnit,
-    });
-    this.recipesService.addIngredient({
-      name: this.enteredIngredients,
-      quanity: this.enteredQuanity,
-      unit:this.enteredUnit,
-    });
-  }
 }
