+++ src/app/recipes/recipe-list/recipe-list.component.ts
@@ -12,21 +12,6 @@ import { RecipeItemComponent } from './recipe-item/recipe-item.component';
 export class RecipeListComponent {
   private recipesService = inject(RecipesService);
   selectedFilter = signal<string>('show-all');
-  // showRecipes = computed(()=>{
-  //   switch(this.selectedFilter()){
-  //     case 'show-favourites':
-  //       return this.recipesService.allRecipes().filter((recipe) => recipe.favourites === true);
-  //     case 'show-non-favourites':
-  //       return this.recipesService.allRecipes().filter((recipes) => recipes.favourites === false);
-  //     default:
-  //       return this.recipesService.allRecipes();
-  //   }
-  // });
-
-  // onChangeFilter(filter: string) {
-  //   this.selectedFilter.set(filter);
-  // }
-
   showRecipesTemp = this.recipesService.allRecipes;
 
   onChangeFilterTemp(filter:string){
