--- src/app/recipes/recipes.component.css
@@ -1,13 +1,0 @@
-.float-parent-element {
-  width: 100%;
-  display: flex;
-  gap: 16px;
-}
-.float-child-element {
-  float: left;
-  width: 50%;
-}
-
-ul {
-  list-style-type: none;
-}
