+++ src/app/recipes/recipes.component.html
@@ -1,5 +1,1 @@
-<div class="float-parent-element">
-  <app-new-recipe (ingredient)="onAddIngredient($event)" (addNewRecipe)="onAddRecipeTest()"class="float-child-element" />
-  <app-ingredients [ingredientsList]="ingredients" class="float-child-element" />
-</div>
 <app-recipe-list />
