+++ src/app/recipes/recipes.component.ts
@@ -1,23 +1,15 @@
-import { Ingredients } from './ingredients/ingredients.model';
+import { Ingredients } from './create-recipe/ingredients/ingredients.model';
 import { Component, input } from '@angular/core';
 import { RecipeListComponent } from './recipe-list/recipe-list.component';
-import { NewRecipeComponent } from './new-recipe/new-recipe.component';
-import { IngredientsComponent } from "./ingredients/ingredients.component";
 
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule} from '@angular/material/icon';
 @Component({
   selector: 'app-recipes',
   standalone: true,
-  imports: [RecipeListComponent, NewRecipeComponent, IngredientsComponent],
+  imports: [RecipeListComponent, MatIconModule, MatButtonModule],
   templateUrl: './recipes.component.html',
   styleUrl: './recipes.component.css'
 })
 export class RecipesComponent {
-  ingredients: Ingredients[] = []
-  onAddIngredient(ingredient: Ingredients){
-    this.ingredients.push(ingredient);
-  }
-
-  onAddRecipeTest(){
-    this.ingredients = [];
-  }
 }
