--- src/styles.css
@@ -1,10 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-
-html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
-
-body {
-  margin: 0;
-  background-color: #2d2b24;
-  color: #f5f1e0;
-}
