+++ src/styles.scss
@@ -1,0 +1,196 @@
+/* These tokens are generated using https://themes.angular-material.dev/ */
+/* Preview: https://themes.angular-material.dev/?bold-font-weight=700&brand-font-family=Roboto&medium-font-weight=500&plain-font-family=Roboto&regular-font-weight=400&seed-error=%23ffb4ab&seed-neutral=%237c776a&seed-neutral-variant=%237e7761&seed-primary=%23f8d63c&seed-secondary=%23feeaa6&seed-tertiary=%23acd45a */
+/* Seed Colors: primary: #f8d63c, secondary: #feeaa6, tertiary: #acd45a, error: #ffb4ab, neutral: #7c776a, neutral-variant: #7e7761 */
+/* Seed Typography: plain-font-family: Roboto, brand-font-family: Roboto, bold-font-weight: 700, medium-font-weight: 500, regular-font-weight: 400 */
+
+/* Make sure to import fonts in `<head>` of html */
+/*
+<style>
+  @import url('https://fonts.googleapis.com/css2?family=Roboto&family=Roboto&display=swap');
+</style>
+
+OR
+
+<link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto&family=Roboto&display=swap">
+*/
+
+@use "@angular/material" as mat;
+
+
+/* Light Theme */
+:root, :host {
+  @include mat.theme-overrides((
+    primary: #6f5d00,
+    on-primary: #ffffff,
+    primary-container: #ffe167,
+    on-primary-container: #221b00,
+    inverse-primary: #e5c52a,
+    primary-fixed: #ffe167,
+    primary-fixed-dim: #e5c52a,
+    on-primary-fixed: #221b00,
+    on-primary-fixed-variant: #544600,
+    secondary: #6b5e28,
+    on-secondary: #ffffff,
+    secondary-container: #f6e29f,
+    on-secondary-container: #221b00,
+    secondary-fixed: #f6e29f,
+    secondary-fixed-dim: #d8c685,
+    on-secondary-fixed: #221b00,
+    on-secondary-fixed-variant: #524612,
+    tertiary: #4a6700,
+    on-tertiary: #ffffff,
+    tertiary-container: #c7f173,
+    on-tertiary-container: #141f00,
+    tertiary-fixed: #c7f173,
+    tertiary-fixed-dim: #acd45a,
+    on-tertiary-fixed: #141f00,
+    on-tertiary-fixed-variant: #374e00,
+    background: #44433f,
+    on-background: #1c1b1a,
+    surface: #fdf8f6,
+    surface-dim: #ddd9d7,
+    surface-bright: #fdf8f6,
+    surface-container-lowest: #ffffff,
+    surface-container-low: #f7f3f0,
+    surface-container: #f1edeb,
+    surface-container-high: #ece7e5,
+    surface-container-highest: #e6e2df,
+    on-surface: #1c1b1a,
+    shadow: #000000,
+    scrim: #000000,
+    surface-tint: #635e52,
+    inverse-surface: #31302f,
+    inverse-on-surface: #f4f0ed,
+    outline: #7b776e,
+    outline-variant: #ccc6ba,
+    neutral: #797775,
+    neutral10: #1c1b1a,
+    error: #ba1a1a,
+    error-container: #ffdad6,
+    on-error: #ffffff,
+    on-error-container: #410002,
+    surface-variant: #e8e2d6,
+    on-surface-variant: #4a473e,
+    neutral-variant: #7b776d,
+    neutral-variant20: #333028,
+    inverse-secondary: #d8c685,
+    inverse-tertiary: #acd45a,
+  ));
+}
+
+/* Dark Theme */
+.dark {
+  @include mat.theme-overrides((
+    primary: #e5c52a,
+    on-primary: #3a3000,
+    primary-container: #544600,
+    on-primary-container: #ffe167,
+    inverse-primary: #6f5d00,
+    primary-fixed: #ffe167,
+    primary-fixed-dim: #e5c52a,
+    on-primary-fixed: #221b00,
+    on-primary-fixed-variant: #544600,
+    secondary: #d8c685,
+    on-secondary: #3a3000,
+    secondary-container: #524612,
+    on-secondary-container: #f6e29f,
+    secondary-fixed: #f6e29f,
+    secondary-fixed-dim: #d8c685,
+    on-secondary-fixed: #221b00,
+    on-secondary-fixed-variant: #524612,
+    tertiary: #acd45a,
+    on-tertiary: #253600,
+    tertiary-container: #374e00,
+    on-tertiary-container: #c7f173,
+    tertiary-fixed: #c7f173,
+    tertiary-fixed-dim: #acd45a,
+    on-tertiary-fixed: #141f00,
+    on-tertiary-fixed-variant: #374e00,
+    background: #141312,
+    on-background: #e6e2df,
+    surface: #141312,
+    surface-dim: #141312,
+    surface-bright: #3a3938,
+    surface-container-lowest: #0f0e0d,
+    surface-container-low: #1c1b1a,
+    surface-container: #201f1e,
+    surface-container-high: #2b2a29,
+    surface-container-highest: #363433,
+    on-surface: #e6e2df,
+    shadow: #000000,
+    scrim: #000000,
+    surface-tint: #cdc6b7,
+    inverse-surface: #e6e2df,
+    inverse-on-surface: #31302f,
+    outline: #959087,
+    outline-variant: #4a473e,
+    neutral: #797775,
+    neutral10: #1c1b1a,
+    error: #ffb4ab,
+    error-container: #93000a,
+    on-error: #690005,
+    on-error-container: #ffdad6,
+    surface-variant: #4a473e,
+    on-surface-variant: #ccc6ba,
+    neutral-variant: #7b776d,
+    neutral-variant20: #333028,
+    inverse-secondary: #6b5e28,
+    inverse-tertiary: #4a6700,
+  ));
+}
+
+/* Typography */
+:root, :host {
+  /* core typography tokens are not generating css variables, hence below */
+  --mat-sys-brand-font-family: Roboto;
+  --mat-sys-plain-font-family: Roboto;
+  --mat-sys-bold-font-weight: 700;
+  --mat-sys-medium-font-weight: 500;
+  --mat-sys-regular-font-weight: 400;
+
+  @include mat.theme-overrides((
+    brand-family: Roboto,
+    plain-family: Roboto,
+    bold-weight: 700,
+    medium-weight: 500,
+    regular-weight: 400,
+  ));
+}
+
+
+:root{
+  color-scheme: dark;
+}
+
+html,
+body {
+  height: 100%;
+  background-color: var(--mat-sys-background);
+}
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+}
+
+//set global styles
+body, img, html, table {
+  border-width: 2px;
+  border-radius: 7px !important;
+}
+
+.card {
+  max-width: 400px;
+  box-shadow: 10px 10px var(--mat-sys-secondary) !important;
+  margin-top: 30px;
+}
+
+.form-shadow-box {
+  margin: 2rem auto;
+  max-width: 30rem;
+  padding: 2rem;
+  border-radius: 4px;
+  background-color: #44433f;
+  border-color: var(--mat-sys-primary-container);
+  box-shadow: 10px 10px var(--mat-sys-primary-container);
+}
+
